rename: 디렉토리 구조 변경, 컴포넌트 및 파일 이름 수정
