+++ src/App.tsx
@@ -1,6 +1,6 @@
 import "./App.css";
 
-import CardEnrollForm from "./components/CardEnrollForm";
+import CardEnrollForm from "./components/payment/CardEnrollForm";
 
 function App() {
   return <CardEnrollForm />;
